pyproject.toml
[tool.poetry]
name = "django-labs-accounts"
version = "0.9.5"
description = "Reusable Django app for Penn Labs accounts"
authors = ["Penn Labs <contact@pennlabs.org>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pennlabs/django-labs-accounts"
repository = "https://github.com/pennlabs/django-labs-accounts"
include = ["CHANGELOG.md", "LICENSE"]
exclude = ["tox.ini"]
packages = [
    { include = "accounts" },
    { include = "identity" },
    { include = "analytics" },
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
requests-oauthlib = "^1.3.1"
requests = "^2.0.0"
djangorestframework = "^3.14.0"
six = "^1.16.0"
jwcrypto = "^1.4.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^3.9.0"
flake8-isort = "^4.0.0"
flake8-quotes = "^3.2.0"
flake8-absolute-import = "^1.0"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
